fix(playground): migrate playground to current charts API (#36)

+++ playground/src/commonMain/kotlin/presentation/chart/renderers/PieChartRenderer.kt
@@ -23,7 +23,7 @@ internal fun PieChartRenderer(spec: ValidatedChartSpec) {
             data.values.mapIndexed { index, value ->
                 PieSlice(
                     label = labels[index],
-                    value = value,
+                    value = value.toDouble(),
                     color = palette?.getOrNull(index)?.toComposeColor(),
                 )
             }
